prepare merge stt button
