+++ app/src/main/AndroidManifest.xml
@@ -14,9 +14,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.NLInterface"
         tools:targetApi="31">
-        <activity
-            android:name=".activities.SpeechToTextActivity"
-            android:exported="false" />
+
         <activity
             android:name=".activities.MainActivity"
             android:exported="true"
