--- app/src/main/java/com/nlinterface/activities/SpeechToTextActivity.kt
@@ -1,40 +1,0 @@
-package com.nlinterface.activities
-
-import android.content.Intent
-import android.os.Bundle
-import android.speech.RecognitionListener
-import android.speech.RecognizerIntent
-import android.speech.SpeechRecognizer
-import android.widget.TextView
-import androidx.appcompat.app.AppCompatActivity
-import com.nlinterface.R
-import com.nlinterface.utility.SpeechToTextButton
-import com.nlinterface.utility.SpeechToTextUtility
-
-class SpeechToTextActivity : AppCompatActivity() {
-
-    private var isListening = false
-    private var outputText: TextView? = null
-    private var sttTrigger: SpeechToTextButton? = null
-    private val speechToTextUtility = SpeechToTextUtility()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_speech_to_text_example)
-
-        outputText = findViewById(R.id.outputTV)
-
-        sttTrigger = findViewById(R.id.stt_btn)
-        sttTrigger!!.setOnClickListener {
-            if (isListening) {
-                speechToTextUtility.handleSpeechEnd(outputText!!, sttTrigger!!)
-                isListening = false
-            } else {
-                speechToTextUtility.handleSpeechBegin(outputText!!, sttTrigger!!)
-                isListening = true
-            }
-        }
-
-        speechToTextUtility.createSpeechRecognizer(this, outputText!!)
-    }
-}
