+++ app/src/main/java/com/nlinterface/utility/SpeechToTextUtility.kt
@@ -7,6 +7,8 @@ import android.speech.RecognitionListener
 import android.speech.RecognizerIntent
 import android.speech.SpeechRecognizer
 import android.widget.TextView
+import androidx.appcompat.app.AppCompatActivity
+import com.nlinterface.R
 
 class SpeechToTextUtility {
     private var speechRecognizer: SpeechRecognizer? = null
@@ -64,3 +66,33 @@ class SpeechToTextUtility {
         return intent
     }
 }
+
+// USE CASE EXAMPLE
+/*
+class SpeechToTextActivity : AppCompatActivity() {
+
+    private var isListening = false
+    private var outputText: TextView? = null
+    private var sttTrigger: SpeechToTextButton? = null
+    private val speechToTextUtility = SpeechToTextUtility()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_speech_to_text_example)
+
+        outputText = findViewById(R.id.outputTV)
+
+        sttTrigger = findViewById(R.id.stt_btn)
+        sttTrigger!!.setOnClickListener {
+            if (isListening) {
+                speechToTextUtility.handleSpeechEnd(outputText!!, sttTrigger!!)
+                isListening = false
+            } else {
+                speechToTextUtility.handleSpeechBegin(outputText!!, sttTrigger!!)
+                isListening = true
+            }
+        }
+
+        speechToTextUtility.createSpeechRecognizer(this, outputText!!)
+    }
+}*/
