+++ app/src/main/res/layout/activity_main.xml
@@ -6,32 +6,6 @@
     android:layout_height="match_parent"
     android:accessibilityTraversalAfter="@id/next_activity">
 
-    <com.nlinterface.utility.SpeechToTextButton
-        android:id="@+id/stt_btn"
-        android:layout_width="386dp"
-        android:layout_height="174dp"
-        android:layout_marginStart="12dp"
-        android:layout_marginEnd="12dp"
-        android:layout_marginBottom="16dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <TextView
-        android:id="@+id/outputTV"
-        android:layout_width="228dp"
-        android:layout_height="46dp"
-        android:layout_marginTop="492dp"
-        android:autoSizeTextType="uniform"
-        android:contentDescription="@string/stt_output_content"
-        android:text="@string/placeholder_text"
-        android:textAlignment="center"
-        android:textSize="24sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.497"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
     <Button
         android:id="@+id/next_activity"
         android:layout_width="300dp"
